Hack around conflict between project.py and static.sh when running project.py --release
In project.py, _copy_files() clones everything, including the lib/ directory.
In static.sh, we try instead to softlink, which fails.

Y'all have been fiddling with this, so please feel free to do a better job of cleaning up than this -- it's just a quick hack to work around breakage at HEAD.
	Change on 2016/01/29 by mgainer <mgainer@google.com>
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=113363020

+++ coursebuilder/scripts/static.sh
@@ -33,7 +33,7 @@ if [ "$ALLOW_STATIC_SERV" = true ] ; then
   if [ ! -d "$LN_TARGET/_static/" ]; then
     mkdir $COURSEBUILDER_RESOURCES/_static/
   fi
-  if [ ! -L "$LN_SOURCE/_static" ]; then
+  if [ ! -L "$LN_SOURCE/_static" -a ! -d "$LN_SOURCE/_static" ]; then
     ln -s $COURSEBUILDER_RESOURCES/_static/ $LN_SOURCE/
   fi
